+++ README.md
@@ -1,8 +1,12 @@
-This workspace contains the crate webpki-roots.
+This workspace contains the crates webpki-roots and webpki-ccadb.
 
 The webpki-roots crate contains Mozilla's root certificates for use with
 the [webpki](https://github.com/rustls/webpki) or
 [rustls](https://github.com/rustls/rustls) crates.
+
+The webpki-ccadb crate populates the root certificates for the webpki-roots crate
+using the data provided by the [Common CA Database (CCADB)](https://www.ccadb.org/).
+Inspired by [certifi.io](https://certifi.io/en/latest/).
 
 The webpki-roots crate is inspired by [certifi.io](https://certifi.io/en/latest/) and
 uses the data provided by the [Common CA Database (CCADB)](https://www.ccadb.org/).
