+++ webpki-ccadb/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "webpki-ccadb"
+version = "0.1.0"
+edition = { workspace = true }
+readme = "README.md"
+license = "MIT OR Apache-2.0"
+homepage = { workspace = true }
+repository = { workspace = true }
+description = "Common CA Database (CCADB) interface for use with webpki-roots"
+
+[dependencies]
+chrono = { version = "0.4.26", default-features = false, features = ["clock"] }
+csv = "1.2.2"
+hex = { workspace = true }
+num-bigint = "0.4.3"
+pki-types = { workspace = true }
+reqwest = { version = "0.11", features = ["rustls-tls-manual-roots"] }
+rustls-pemfile = "2.0.0"
+serde = { version = "1.0.183", features = ["derive"] }
+webpki = { workspace = true }
+x509-parser = { workspace = true }
+yasna = { workspace = true }
