+++ webpki-roots/Cargo.toml
@@ -1,29 +1,24 @@
 [package]
 name = "webpki-roots"
 version = "0.26.0"
-edition = "2018"
+edition = { workspace = true }
 readme = "README.md"
 license = "MPL-2.0"
-homepage = "https://github.com/rustls/webpki-roots"
-repository = "https://github.com/rustls/webpki-roots"
+homepage = { workspace = true }
+repository = { workspace = true }
 description = "Mozilla's CA root certificates for use with webpki"
 
 [dependencies]
-pki-types = { package = "rustls-pki-types", version = "1", default-features = false }
+pki-types = { workspace = true }
 
 [dev-dependencies]
-chrono = { version = "0.4.26", default-features = false, features = ["clock"] }
-csv = "1.2.2"
-hex = "0.4.3"
-num-bigint = "0.4.3"
+hex = { workspace = true }
 percent-encoding = "2.3"
 rcgen = "0.11.1"
-reqwest = { version = "0.11", features = ["rustls-tls-manual-roots"] }
 ring = "0.17.0"
-rustls-pemfile = "2.0.0"
 rustls = "0.22"
-serde = { version = "1.0.183", features = ["derive"] }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
-webpki = { package = "rustls-webpki", version = "0.102", features = ["alloc"] }
-x509-parser = "0.15.1"
-yasna = "0.5.2"
+webpki = { workspace = true }
+webpki-ccadb = { path = "../webpki-ccadb" }
+x509-parser = { workspace = true }
+yasna = { workspace = true }
